refactor(Balance): AND-1216 Introduce Precision to cope with the two Eth string formats

+++ app/src/main/java/piuk/blockchain/android/ui/account/AccountPresenter.kt
@@ -498,7 +498,7 @@ class AccountPresenter @Inject internal constructor(
 
     private fun getUiString(amount: Long): String {
         return if (currencyState.isDisplayingCryptoCurrency) {
-            currencyFormatManager.getFormattedSelectedCoinValueWithUnit(amount.toBigDecimal())
+            currencyFormatManager.getFormattedSelectedCoinValueWithUnit(amount.toBigInteger())
         } else {
             currencyFormatManager.getFormattedFiatValueFromSelectedCoinValueWithSymbol(amount.toBigDecimal())
         }
